requirements.txt
Flask==2.2.2
tensorflow==2.10.0
numpy==1.23.5
Pillow==9.3.0
transformers==4.24.0
torch==1.13.1
scikit-learn==1.1.3
gunicorn==20.1.0